Delete paper_data directory (still accessible under zenodo)
When working on the revision, we adapted the diffMONT script. And rerun the paper analyses. To not generate confusion, the result files used for the publication were removed from github. They are still accessible in zenodo. The manuscript links to the zenodo repository.

This github repository now only has a minimal data example for testing the code. It is strongly subsampled, so that diffMONT can be tested easily in a few minutes.